Refactored Markdown filter.
