+++ ipymd/formats/python.py
@@ -11,6 +11,7 @@ import ast
 from collections import OrderedDict
 
 from ..lib.base_lexer import BaseGrammar, BaseLexer
+from ..lib.markdown import MarkdownFilter
 from ..ext.six import StringIO
 from ..utils.utils import _ensure_string, _preprocess
 
@@ -128,25 +129,6 @@ def _add_hash(source):
     return source
 
 
-def _replace_header_filter(filter):
-    return {'h1': '# ',
-            'h2': '## ',
-            'h3': '### ',
-            'h4': '#### ',
-            'h5': '##### ',
-            'h6': '###### ',
-            }[filter]
-
-
-def _filter_markdown(source, filters):
-    """Only keep some Markdown headers from a Markdown string."""
-    lines = source.splitlines()
-    # Filters is a list of 'hN' strings where 1 <= N <= 6.
-    headers = [_replace_header_filter(filter) for filter in filters]
-    lines = [line for line in lines if line.startswith(tuple(headers))]
-    return '\n'.join(lines)
-
-
 class PythonReader(object):
     """Python reader."""
     def read(self, python):
@@ -168,45 +150,21 @@ class PythonReader(object):
 
 
 class PythonWriter(object):
-    """Python writer.
-
-    Parameters
-    ----------
-
-    keep_markdown : str | None or False
-        What to keep from Markdown cells. Can be:
-
-        * None or False: don't keep Markdown contents
-        * 'all': keep all Markdown contents
-        * 'headers': just keep Markdown headers
-        * 'h1,h3': just keep headers of level 1 and 3 (can be any combination)
-
-    """
-
+    """Python writer."""
     def __init__(self, keep_markdown='all'):
         self._output = StringIO()
-        if keep_markdown == 'headers':
-            keep_markdown = 'h1,h2,h3,h4,h5,h6'
-        self._keep_markdown = keep_markdown
+        self._markdown_filter = MarkdownFilter(keep_markdown)
 
     def _new_paragraph(self):
         self._output.write('\n\n')
 
     def append_comments(self, source):
         source = source.rstrip()
-
-        # Skip Markdown cell.
-        if not self._keep_markdown:
-            return
-        # Only keep some Markdown headers if keep_markdown is not 'all'.
-        elif self._keep_markdown != 'all':
-            to_keep = self._keep_markdown.split(',')
-            source = _filter_markdown(source, to_keep)
-
+        # Filter Markdown contents.
+        source = self._markdown_filter(source)
         # Skip empty cells.
         if not source:
             return
-
         comments = _add_hash(source)
         self._output.write(comments)
         self._new_paragraph()
